Fix bug with parsing binary files
A problem was introduced in Airflow 2.6.0. Due to [PR30495](https://github.com/apache/airflow/pull/30495)
in Airflow, if any binary files happen to be in the `dags` folder,
the dag processor manager will constantly crash, resulting in
no DAGs being properly parsed. It's caused by some over-eager
parsing of all files, regardless of type.

The working fix is to ensure that these binary files are in the .airflowignore
file, as mentioned [here](https://github.com/apache/airflow/discussions/31519).

A different fix was supposed to land in 2.6.1, but having tested this,
it does not seem to work (yet). (https://github.com/apache/airflow/pull/31401)