optionally embed OpenJDK in whirl image
Iff the `OPENJDK_VERSION` is set in `.whirl.env`, then OpenJDK of that version
is installed during the docker build, and the docker image tag includes the JRE
version number.

To make use of this, you have to add the `OPENJDK_VERSION` to your .whirl.env
file, and don't forget to update the image tag in the `docker-compose` file for
the environment.